fix(profile): addressing comments
